Web application update service references
fix: Use the new calculation service
