+++ TradeCommissionApp.AppHost/Program.cs
@@ -4,11 +4,13 @@ var cache = builder.AddRedis("cache");
 
 var apiService = builder.AddProject<Projects.TradeCommissionApp_ApiService>("apiservice");
 
-builder.AddProject<Projects.TradeCommissionApp_Web>("webfrontend")
-    .WithReference(cache)
+var calculationService = builder.AddProject<Projects.TradeCommissionApp_CalculationService>("calculationservice")
     .WithReference(apiService);
 
-builder.AddProject<Projects.TradeCommissionApp_CalculationService>("calculationservice")
-    .WithReference(apiService);
+builder.AddProject<Projects.TradeCommissionApp_Web>("webfrontend")
+    .WithReference(cache)
+    .WithReference(apiService)
+    .WithReference(calculationService);
+
 
 builder.Build().Run();
